refactor(speech): remove chatterbox TTS
Installed 2026-07-10, reports ready, 9.5GB on disk — a 5.7GB venv plus 3.8GB of
weights — and it has synthesized nothing. The voices, uploads and outputs
directories are all still empty, the worker is stopped, and voice_count is 0.
Chatterbox cannot speak without a voice profile and none was ever recorded.

modules/speech was the largest module in the controller (3,376 lines) for a
feature with no recorded use. Going with it: the audio routes, the stt and tts
services, the speech contract, the desktop plugin bundle, the chatterbox voice
UI, and the speech API client. The plugin runtime's hostCapability mechanism
goes too — it was hardcoded to the chatterbox-voice plugin id.

Two things fell out as dead once speech left, both found by knip rather than by
me: nvidia-compute-processes.ts, and boundedFormData, whose only caller was
multipart audio upload.

The GPU lease registry also went. It had two owners, "llm" and "speech"; the
llm owner only validated against the instance record it already had, so the
speech worker's pinned hold was the entire reason it existed. What survives is
recipe GPU-visibility resolution, which is what compute/bridge.ts actually
uses, so gpu-leases.ts becomes gpu-visibility.ts.

Speech-to-text is unaffected — it moved to the local machine one commit ago.
The 9.5GB on the controller is untouched; deleting it is a separate call.
