+++ frontend/src/app/api/proxy/[...path]/proxy-timeouts.ts
@@ -4,15 +4,8 @@ const SYSTEM_UPSTREAM_TIMEOUT_MS = 20_000;
 const CHAT_COMPLETION_UPSTREAM_TIMEOUT_MS = 600_000;
 const MODEL_LIFECYCLE_TIMEOUT_MS = 360_000;
 const SSE_CONNECT_TIMEOUT_MS = 5_000;
-const SPEECH_GENERATION_TIMEOUT_MS = 360_000;
-const VOICE_REFERENCE_TIMEOUT_MS = 120_000;
 const POST_TIMEOUTS = new Map([
   ["studio/downloads", DOWNLOAD_UPSTREAM_TIMEOUT_MS],
-  ["v1/audio/install", SYSTEM_UPSTREAM_TIMEOUT_MS],
-  ["v1/audio/install/cancel", SYSTEM_UPSTREAM_TIMEOUT_MS],
-  ["v1/audio/speech", SPEECH_GENERATION_TIMEOUT_MS],
-  ["v1/audio/voices", VOICE_REFERENCE_TIMEOUT_MS],
-  ["v1/audio/runtime/stop", SYSTEM_UPSTREAM_TIMEOUT_MS],
 ]);
 
 export function getUpstreamTimeoutMs(path: string[], method = "GET"): number {
