+++ frontend/src/lib/api/create-api-client.ts
@@ -3,7 +3,6 @@ import { createApiCore } from "./core";
 import { createLogsApi } from "./logs";
 import { createRecipesApi } from "./recipes";
 import { createStudioApi } from "./studio";
-import { createSpeechApi } from "./speech";
 import { createSystemApi } from "./system";
 
 export function createApiClient(params: {
@@ -19,7 +18,6 @@ export function createApiClient(params: {
     ...createLogsApi(core),
     ...createStudioApi(core),
     ...createRigsApi(core),
-    ...createSpeechApi(core),
     healthPoll: (timeoutMs?: number) => core.healthPoll(timeoutMs),
   };
 }
